Roll src/net/third_party/quiche/src/ 1125f4e47..899981cad4c3 (10 commits)
https://quiche.googlesource.com/quiche.git/+log/1125f4e47606..899981cad4c3

$ git log 1125f4e47..899981cad --date=short --no-merges --format='%ad %ae %s'
2023-09-14 wub No public description
2023-09-14 wub No public description
2023-09-13 birenroy Moves metadata-related tests to a separate test file.
2023-09-13 vasilvv Deflake quic_generic_session_test
2023-09-13 birenroy Adds an OgHttp2Session option for the initial assumed value of the remote endpoint's max concurrent streams setting.
2023-09-13 quiche-dev Add missing absl includes
2023-09-13 quiche-dev Allow truncating of binaryhttp request
2023-09-12 vasilvv Set a fixed CC algorithm in quic_generic_session_test
2023-09-11 bnc Make QuicErrorCode underlying type uint32_t, avoid narrowing when parsing.
2023-09-11 quiche-dev Registering a new use case to test out go/at-integration-guide.

Created with:
  roll-dep src/net/third_party/quiche/src

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4864769
Commit-Queue: Victor Vasiliev <vasilvv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197350}

+++ net/quic/dedicated_web_transport_http3_client.cc
@@ -126,10 +126,11 @@ void RecordNetLogQuicSessionClientStateChanged(
                         .Set("last_state", WebTransportStateString(last_state))
                         .Set("next_state", WebTransportStateString(next_state));
         if (error.has_value()) {
-          dict.Set("error", base::Value::Dict()
-                                .Set("net_error", error->net_error)
-                                .Set("quic_error", error->quic_error)
-                                .Set("details", error->details));
+          dict.Set("error",
+                   base::Value::Dict()
+                       .Set("net_error", error->net_error)
+                       .Set("quic_error", static_cast<int>(error->quic_error))
+                       .Set("details", error->details));
         }
         return dict;
       });
